eapp updates [Shinomori]
git-svn-id: http://eathena.googlecode.com/svn/branches/eapp@2390 9c4de8d1-4af9-ca9b-867f-b47ac08f6c80

+++ src/map/pc.h
@@ -43,7 +43,7 @@ static inline bool pc_breakhelm(struct map_session_data &sd)	{ return pc_break_e
 
 int pc_checkskill(struct map_session_data &sd,unsigned short skill_id);
 int pc_checkallowskill(struct map_session_data &sd);
-int pc_checkequip(struct map_session_data &sd,int pos);
+unsigned short pc_checkequip(struct map_session_data &sd, unsigned short pos);
 
 int pc_calc_skilltree(struct map_session_data &sd);
 int pc_calc_skilltree_normalize_job(struct map_session_data &sd, int c);
